backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "AI-Powered Todo Chatbot Backend"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
alembic = ">=1.18.2"
fastapi = ">=0.128.0"
fastmcp = ">=2.14.4"
httpx = ">=0.28.1"
openai-agents = ">=0.7.0"
passlib = ">=1.7.4"
psycopg2-binary = ">=2.9.11"
pydantic-settings = ">=2.12.0"
python-jose = ">=3.5.0"
python-multipart = ">=0.0.22"
sqlmodel = ">=0.0.31"
uvicorn = ">=0.40.0"
bcrypt = "<4.0.0"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^7.0.0"
pytest-asyncio = "^1.3.0"

